requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
exceptiongroup==1.3.0
fastapi==0.119.0
h11==0.16.0
idna==3.11
msal==1.34.0
pycparser==2.23
pydantic==2.12.2
pydantic_core==2.41.4
PyJWT==2.10.1
python-dotenv==1.1.1
requests==2.32.5
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
